Make cross site linking simpler
Before, it used login_url to go to the login page which was
prepopulated with the current user's name. This doesn't work if you use
a non-standard authentication system. It should respect
Alchemy.login_path but even this isn't really working since the form's
fields could have different names than the ones used here. So better
just link to the dashboard (the site you're currently on) and let
alchemy redirect to the login page. We lose the user name prepopulation
but also lose the additional step of going to the login page if the
user is logged in already.