Update whole element upon save in editor
Before, only the element editor's head was replaced. This may lead to
problems when the editor body or one of it's essences' editors doesn't
just display inputs. We had one problem where we displayed an image of
a referenced model. This was never updated. Now it gets updated since
the whole editor block will be regenerated.